bugfix: stat in copier is now size_t

+++ copier.h
@@ -1,7 +1,7 @@
 #ifndef __COPIER__
 #define __COPIER__
 
-int copy_step(subf_t*, int, int, int);
+size_t copy_step(subf_t*, int, int, int);
 
 #endif
